refactor: Remove legacy configuration tests and streamline settings loading
- Eliminate outdated tests from `system_tests.rs` related to configuration management.
- Update the `load` method in the `Settings` trait to ensure default configurations are returned when the specified file does not exist.
- Introduce new tests in `macro_tests.rs` to validate the loading and saving of configuration settings, including handling non-existent files.
