+++ utils/settings_derive/Cargo.toml
@@ -10,3 +10,13 @@ proc-macro = true
 syn = "1.0"
 quote = "1.0"
 proc-macro2 = "1.0"
+
+[dev-dependencies]
+tempfile = "3.10.1"
+assert_cmd = "2.0"
+serde.workspace = true
+serde_yaml = "0.9.33"
+once_cell = "1.19.0"
+anyhow.workspace = true
+settings = {path = "../settings", features = ["derive"]}
+settings_derive = {path = "."}
